Renamed stream_pcs to audio_pcs
